+++ services/vmwarecloudsimple/mgmt/2019-04-01/vmwarecloudsimple/CHANGELOG.md
@@ -1,2 +1,9 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Struct Fields
+
+1. DedicatedCloudNodeProperties.Created changed type from interface{} to *date.Time